requirements.txt
asgiref==3.2.10
Django==3.1.2
mysql-connector-python==8.0.21
protobuf==3.13.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
